Bugfixes to fgbg; general improvments to *.bash

+++ bin/killtree.bash
@@ -6,14 +6,19 @@
 # Should also work on POSIX.
 # The first arg is the pid of the tree to kill. The second arg is the kill signal
 # (term is default).
+# If process doesn't exist, terminates with status 0.
 
 killtree() {
 	local _pid=$1
+	[[ ! $(ps -o "pid=" | grep -x $_pid) ]] && return 0
 	local _sig=${2-TERM}
+
+	# Snag the child and killtree it
 	local _regex="[ ]*([0-9]+)[ ]+${_pid}"
 	for _child in $(ps ax -o "pid= ppid=" | grep -E "${_regex}" | sed -E "s/${_regex}/\1/g"); do
 		killtree ${_child} ${_sig}
 	done
+
 	kill -${_sig} ${_pid}
 }
